:rocket: Moved identifiers service from local memorandum-provider to shared-identifiers
