+++ backend/libs/memorandum/memorandum-controller/src/lib/memorandum-controller.module.ts
@@ -1,4 +1,4 @@
-import { MemorandumProviderModule } from '@backend/memorandum/memorandum-provider';
+import { IdentifiersModule } from '@backend/shared-identifiers';
 import { SharedKeystorePersistenceModule } from '@backend/shared-keystore-persistence';
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
@@ -6,11 +6,7 @@ import { IdentifiersController } from './identifiers.controller';
 import { KeystoreController } from './keystore.controller';
 
 @Module({
-  imports: [
-    MemorandumProviderModule,
-    SharedKeystorePersistenceModule,
-    ConfigModule,
-  ],
+  imports: [IdentifiersModule, SharedKeystorePersistenceModule, ConfigModule],
   controllers: [IdentifiersController, KeystoreController],
   providers: [],
   exports: [],
