--- backend/libs/memorandum/memorandum-provider/README.md
@@ -1,7 +1,0 @@
-# memorandum-memorandum-provider
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test memorandum-memorandum-provider` to execute the unit tests via [Jest](https://jestjs.io).
