--- backend/libs/memorandum/memorandum-provider/project.json
@@ -1,20 +1,0 @@
-{
-  "name": "memorandum-memorandum-provider",
-  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/memorandum/memorandum-provider/src",
-  "projectType": "library",
-  "tags": [],
-  "targets": {
-    "lint": {
-      "executor": "@nx/eslint:lint",
-      "outputs": ["{options.outputFile}"]
-    },
-    "test": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "libs/memorandum/memorandum-provider/jest.config.ts"
-      }
-    }
-  }
-}
