--- backend/libs/memorandum/memorandum-provider/src/index.ts
@@ -1,2 +1,0 @@
-export * from './lib/identifier.service';
-export * from './lib/memorandum-provider.module';
