--- backend/libs/memorandum/memorandum-provider/src/lib/memorandum-provider.module.ts
@@ -1,11 +1,0 @@
-import { MemorandumPersistenceModule } from '@backend/memorandum/memorandum-persistence';
-import { Module } from '@nestjs/common';
-import { IdentifiersService } from './identifier.service';
-
-@Module({
-  imports: [MemorandumPersistenceModule],
-  controllers: [],
-  providers: [IdentifiersService],
-  exports: [IdentifiersService],
-})
-export class MemorandumProviderModule {}
