+++ backend/libs/shared/provider/identifiers/src/index.ts
@@ -1,1 +1,2 @@
+export * from './lib/identifier.service';
 export * from './lib/identifiers.module';
