+++ backend/libs/shared/provider/identifiers/src/lib/identifiers.module.ts
@@ -1,8 +1,11 @@
+import { MemorandumPersistenceModule } from '@backend/memorandum/memorandum-persistence';
 import { Module } from '@nestjs/common';
+import { IdentifiersService } from './identifier.service';
 
 @Module({
+  imports: [MemorandumPersistenceModule],
   controllers: [],
-  providers: [],
-  exports: [],
+  providers: [IdentifiersService],
+  exports: [IdentifiersService],
 })
 export class IdentifiersModule {}
